add permission checks to event commands

+++ lib/rubybot/modules/event_commands.rb
@@ -9,7 +9,12 @@ module RubyBot
 
       require_relative '../data/db'
 
-      command(:set_log_channels, description: 'Sets the channel the bot logs to', help_available: false) do |event, *channels|
+      command(:set_log_channels,
+              description: 'Sets the channel the bot logs to',
+              required_permissions: [:manage_messages],
+              help_available: false) do |event, *channels|
+        break unless event.user.id == RubyBot.config['admin_id']
+
         channels.each do |channel|
           RubyBot::DB.save_log_channel(event.server.id, channel)
         end
@@ -19,7 +24,12 @@ module RubyBot
         end
       end
 
-      command(:list_log_channels, description: 'Lists log channels configured for this server', help_available: false) do |event|
+      command(:list_log_channels,
+              description: 'Lists log channels configured for this server',
+              required_permissions: [:manage_messages],
+              help_available: false) do |event|
+        break unless event.user.id == RubyBot.config['admin_id']
+
         channels = RubyBot::DB.find_log_channels(event.server.id).map(&:channel_id)
         event.send_embed do |embed|
           embed.color = '#00ff00'
@@ -30,6 +40,8 @@ module RubyBot
 
       command(:remove_log_channels,
               description: 'Removes log channels configured for this server', help_available: false) do |event, *channels|
+        break unless event.user.id == RubyBot.config['admin_id']
+
         channels.each do |channel|
           RubyBot::DB.remove_log_channel(event.server.id, channel)
         end
@@ -39,7 +51,12 @@ module RubyBot
         end
       end
 
-      command(:set_log_blacklist_channels, description: 'Blacklists a channel from logging', help_available: false) do |event, *channels|
+      command(:set_log_blacklist_channels,
+              description: 'Blacklists a channel from logging',
+              required_permissions: [:manage_messages],
+              help_available: false) do |event, *channels|
+        break unless event.user.id == RubyBot.config['admin_id']
+
         channels.each do |channel|
           RubyBot::DB.save_log_blacklist_channel(event.server.id, channel)
         end
@@ -49,7 +66,12 @@ module RubyBot
         end
       end
 
-      command(:list_log_blacklist_channels, description: 'Lists all blacklisted channels', help_available: false) do |event|
+      command(:list_log_blacklist_channels,
+              description: 'Lists all blacklisted channels',
+              required_permissions: [:manage_messages],
+              help_available: false) do |event|
+        break unless event.user.id == RubyBot.config['admin_id']
+
         channels = RubyBot::DB.find_log_blacklist_channels(event.server.id).map(&:channel_id)
         event.send_embed do |embed|
           embed.color = '#00ff00'
@@ -58,7 +80,12 @@ module RubyBot
         end
       end
 
-      command(:remove_log_blacklist_channels, description: 'Removes log channels configured for this server', help_available: false) do |event, *channels|
+      command(:remove_log_blacklist_channels,
+              description: 'Removes log channels configured for this server',
+              required_permissions: [:manage_messages],
+              help_available: false) do |event, *channels|
+        break unless event.user.id == RubyBot.config['admin_id']
+
         channels.each do |channel|
           RubyBot::DB.remove_log_blacklist_channel(event.server.id, channel)
         end
@@ -68,7 +95,9 @@ module RubyBot
         end
       end
 
-      command(:blacklist_string, description: 'Marks messages containing <String> for deletion') do |event, *content|
+      command(:blacklist_string,
+              description: 'Marks messages containing <String> for deletion',
+              required_permissions: [:manage_messages]) do |event, *content|
         content.each do |text|
           RubyBot::DB.save_forbidden_strings(event.server.id, text)
         end
@@ -78,7 +107,7 @@ module RubyBot
         end
       end
 
-      command(:show_blacklist, description: 'Lists all forbidden Strings') do |event|
+      command(:show_blacklist, description: 'Lists all forbidden Strings', required_permissions: [:manage_messages]) do |event|
         content = RubyBot::DB.find_forbidden_strings(event.server.id).map(&:text)
         event.send_embed do |embed|
           embed.color = '#00ff00'
